포스트 삭제시 댓글과 좋아요 삭제되도록 구현

+++ src/main/java/com/practice/sns/service/LikeService.java
@@ -54,7 +54,7 @@ public class LikeService {
 
     }
 
-    public int countLike(Long postId) {
+    public long countLike(Long postId) {
 
         // 포스트를 찾는다
         Post post = postRepository.findById(postId)
